moved merge json
